fix: use postgres user for kong-database healthcheck
This following log was a hint that some recurrent task would not use an
existing PostgreSQL user:

    kong-database_1   | FATAL:  role "root" does not exist

The proposed fix ensures `healthcheck` is always using the default
`postgres` user instead of `$(whoami)`.

We switch `CMD-SHELL` to `CMD` so we can pass arguments to `pg_isready`
in subsequent elements of the Array.

It seems the shell does not inherit the environment variables passed to
the `environment` attribute, so using `$POSTGRES_USER` did not work in
my initial tests.