feat(entrypoint) allow users to configure uid/gid of kong user
If `KONG_UID` or `KONG_GID` environment variable is set, change the uid
and gid of the kong user (and group if it exists).

On alpine, kong group doesn't exist, so that is not allowed.

This change makes Kong compatible with Istio.

From https://istio.io/docs/setup/kubernetes/additional-setup/requirements/:

> Application UIDs: Ensure your pods do not run applications as a user
with the user ID (UID) value of 1337.