chore(service mesh) service mesh is no longer part of kong 2.0 (#308)
* fix(openshift) allow images to run as user kong

Per the OpenShift docs [1] if the Dockerfile does not have the USER
directive, it will be run as the root UID, which was the case for our
alpine and centos images. The problem is that running as root on some
OpenShift setups (e.g. Minishift) did not allow Kong to be started:

```
~/Downloads$ oc logs -f deployment/kong-rc
chmod: /proc/self/fd/1: Permission denied
chmod: /proc/self/fd/2: Permission denied
nginx: [emerg] open() "/dev/stderr" failed (13: Permission denied)
```

This fixes an OpenShift issue where images were not running as root
because no USER was specified in the alpine and centos dockerfiles.

[1]
https://docs.openshift.com/enterprise/3.2/admin_guide/manage_scc.html#enable-images-to-run-with-user-in-the-dockerfile

* chore(*) remove service mesh support

Removing `setcap` and `su-exec` related changes.

Co-authored-by: Murillo Paula <murillo@murillopaula.com>