fix(update.sh) automatize ubuntu SHA setting
The ubuntu download package contains the ubuntu codename (`focal`, at the moment). When calculating the url to download, the code name was evaluated inline by `$(cat /etc/os-release ...)`

The update script (update.sh) is not able to read that url with the inline calculation using the simple regex it uses for all the other distros. As a result, the ubuntu SHA was not updated by the update.sh script, that step was simply skipped. This was manually fixed means that on every release, the build would error, the person doing the release would notice the problem, manually add the ubuntu sha, and continue the build.

This change moves the calculation of the codename to a different line and reactivates the calculation of the SHA, so that manual step is no longer necessary.